+++ tests/FullApplicationTest.php
@@ -3,6 +3,7 @@
 use Mockery as m;
 use Illuminate\Http\Request;
 use Laravel\Lumen\Application;
+use Symfony\Component\HttpFoundation\Request as SymfonyRequest;
 
 class FullApplicationTest extends PHPUnit_Framework_TestCase
 {
@@ -22,6 +23,18 @@ class FullApplicationTest extends PHPUnit_Framework_TestCase
         $response = $app->handle(Request::create('/', 'GET'));
         $this->assertEquals(200, $response->getStatusCode());
         $this->assertEquals('Hello World', $response->getContent());
+    }
+
+    public function testBasicSymfonyRequest()
+    {
+        $app = new Application;
+
+        $app->get('/', function () {
+            return response('Hello World');
+        });
+
+        $response = $app->handle(SymfonyRequest::create('/', 'GET'));
+        $this->assertEquals(200, $response->getStatusCode());
     }
 
     public function testAddRouteMultipleMethodRequest()